requirements.txt
    annotated-types==0.7.0
    anyio==4.10.0
    blinker==1.9.0
    cachetools==5.5.2
    certifi==2025.8.3
    charset-normalizer==3.4.3
    click==8.2.1
    colorama==0.4.6
    distro==1.9.0
    filetype==1.2.0
    Flask==3.1.1
    flask-cors==6.0.1
    google-ai-generativelanguage==0.6.18
    google-api-core==2.25.1
    google-auth==2.40.3
    googleapis-common-protos==1.70.0
    greenlet==3.2.4
    grpcio==1.74.0
    grpcio-status==1.74.0
    h11==0.16.0
    httpcore==1.0.9
    httpx==0.28.1
    idna==3.10
    itsdangerous==2.2.0
    Jinja2==3.1.6
    jiter==0.10.0
    jsonpatch==1.33
    jsonpointer==3.0.0
    langchain==0.3.27
    langchain-core==0.3.74
    langchain-google-genai==2.1.9
    langchain-openai==0.3.30
    langchain-text-splitters==0.3.9
    langsmith==0.4.14
    MarkupSafe==3.0.2
    openai==1.99.9
    orjson==3.11.2
    packaging==25.0
    proto-plus==1.26.1
    protobuf==6.31.1
    pyasn1==0.6.1
    pyasn1_modules==0.4.2
    pydantic==2.11.7
    pydantic_core==2.33.2
    python-dotenv==1.1.1
    PyYAML==6.0.2
    regex==2025.7.34
    requests==2.32.4
    requests-toolbelt==1.0.0
    rsa==4.9.1
    sniffio==1.3.1
    SQLAlchemy==2.0.43
    tenacity==9.1.2
    tiktoken==0.11.0
    tqdm==4.67.1
    typing-inspection==0.4.1
    typing_extensions==4.14.1
    urllib3==2.5.0
    Werkzeug==3.1.3
    zstandard==0.23.0
